+++ src/Component/JsonSchema/Tests/fixtures/multi-namespace/expected/Schema3/Normalizer/JaneObjectNormalizer.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Jane\Component\JsonSchema\Tests\Expected\Schema2\Normalizer;
+namespace Jane\Component\JsonSchema\Tests\Expected\Schema3\Normalizer;
 
-use Jane\Component\JsonSchema\Tests\Expected\Schema2\Runtime\Normalizer\CheckArray;
-use Jane\Component\JsonSchema\Tests\Expected\Schema2\Runtime\Normalizer\ValidatorTrait;
+use Jane\Component\JsonSchema\Tests\Expected\Schema3\Runtime\Normalizer\CheckArray;
+use Jane\Component\JsonSchema\Tests\Expected\Schema3\Runtime\Normalizer\ValidatorTrait;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareTrait;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
@@ -17,8 +17,6 @@ class JaneObjectNormalizer implements DenormalizerInterface, NormalizerInterface
     use CheckArray;
     use ValidatorTrait;
     protected $normalizers = [
-        
-        \Jane\Component\JsonSchema\Tests\Expected\Schema2\Model\Foo::class => \Jane\Component\JsonSchema\Tests\Expected\Schema2\Normalizer\FooNormalizer::class,
         
         \Jane\Component\JsonSchema\Tests\Expected\Schema3\Model\Bar::class => \Jane\Component\JsonSchema\Tests\Expected\Schema3\Normalizer\BarNormalizer::class,
     ], $normalizersCache = [];
@@ -58,7 +56,6 @@ class JaneObjectNormalizer implements DenormalizerInterface, NormalizerInterface
     {
         return [
             
-            \Jane\Component\JsonSchema\Tests\Expected\Schema2\Model\Foo::class => false,
             \Jane\Component\JsonSchema\Tests\Expected\Schema3\Model\Bar::class => false,
         ];
     }